fix(frame): size the pool request in int so a maximal payload length cannot wrap
AllocPayload asked the pool for uint32(n + Headroom). A header that
announces a payload of 0xFFFFFFFF bytes, which any peer can send, made
that wrap to 51, and slicing the 4 KB buffer to the announced length
panicked in the receive path. The base returned a read error for the
same input. bpool.Get now takes an int, and the request is computed in
int all the way. Regression test added, skipped on Windows because the
announced length is reserved before the read fails.

Also covered: Wire refuses a header longer than Headroom, and Reset
returns the buffer that a From frame grew into.

+++ internal/send.go
@@ -48,7 +48,7 @@ func SendFrame(w io.Writer, fr *frame.Frame) error {
 		return writeVector(w, h, p)
 	}
 
-	pb := bpool.Get(uint32(n))
+	pb := bpool.Get(n)
 	buf := (*pb)[:0]
 	buf = append(buf, h...)
 	buf = append(buf, p...)
